BigQuery INTEGER should be Long in scala

+++ scio-bigquery/src/main/scala/com/spotify/scio/bigquery/types/SchemaUtil.scala
@@ -34,7 +34,7 @@ object SchemaUtil {
 
   private def getRawType(tfs: TableFieldSchema, indent: Int): (String, Seq[String]) = {
     val name = tfs.getType match {
-      case "INTEGER" => "Int"
+      case "INTEGER" => "Long"
       case "FLOAT" => "Double"
       case "BOOLEAN" => "Boolean"
       case "STRING" => "String"
